Make filter and reject work with shell functions/aliases
The `xargs` variants wouldn't work with shell functions and aliases, because `xargs` runs in a separate process from the shell, and the commands it issues run in yet another process. So the shell functions and aliases don't propagate down into those processes. Unfortunately, `xargs` is a *lot* faster.

I suppose we could decide whether to use `xargs` or `IFS` and `for`, by running `type -P` against `$1`. Pull requests accepted.